// src/app/models/students.model.ts
export interface Students {
    sasId: number;
    assessementId: number;
    schoolId: string;
    studentName: string;
    missingAnrolmentNo: boolean;
    primaryLanguage: string;
    lastYearImprovedFlag: boolean;
    programType: string;
    serviceMinutes: string;
    yearsOfForeignEducation: number;
    screeningType: string;
    screeningScore: string;
    comment: string;

}